create new feature category and create css for categories component , added function on request.js
