+++ client/src/main.jsx
@@ -4,18 +4,20 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import App from "./App";
 import Accueil from "./pages/Accueil";
 import Catalogue from "./pages/Catalogue";
+import Category from "./pages/Category";
 import Categories from "./pages/Categories";
 import Favoris from "./pages/Favoris";
 import Basket from "./pages/Basket";
 import About from "./pages/About";
 import GamePage from "./pages/GamePage";
 
-import { fetchData, fetchGameInfo } from "./services/request";
+import { fetchData, fetchGameInfo, fetchCategories } from "./services/request";
 
 import "./styles/app.css";
 import "./styles/gamespages.css";
 import "./styles/navbar.css";
 import "./styles/categorieitem.css";
+import "./styles/categories.css";
 
 const router = createBrowserRouter([
   {
@@ -32,7 +34,12 @@ const router = createBrowserRouter([
       {
         path: "/categories",
         element: <Categories />,
-        loader: () => fetchData().then((data) => data),
+        loader: fetchCategories,
+      },
+      {
+        path: "/category",
+        element: <Category />,
+        loader: fetchData,
       },
       {
         path: "/favoris",
